requirements.txt
requests~=2.31.0
openai~=0.27.8
colorama~=0.4.6
python-dotenv~=1.0.0
termcolor~=2.3.0
tqdm~=4.65.0
google-serp-api
google-search-results
langchain